completed Cart management endpoints :i
:i

Completed Cart management endpoints
the most troublesome problem is trying to show only the really independent data fields for input in the BrowsableAPIView HTML form--menuitem and quantity, not user, unit_price, and price. Tried a few ways to work around, finally decided to remove unit_price and price from Cart model, since they are not really independent from menuitem and quantity. SerializerMethodField and the corresponding get_fieldname functions are really used for retrieving, not creating, data from/to the database.
I had to leave the 'user' field in the serializer and the form. No way without some dirty hacking seems bypass a field in serializer from showing in the form but still could keep it talking to the model. I bypassed 'user' in the overriding create() function in CartCustomerView class, so that the database will always record 'user' as the user of request, not the username entered into the form. It may be confusing to have a dummy field in the form, but this is just in the API view. Technically, this dummy field should better be hidden in front-end work.